add bashInteractive to nix env to fix mac flake
on mac, bash_kernel fails with the bare env from a timeout when the
client tries to connect to the kernel. reason is unclear, but since
nix-shell still worked as expected, comparing the bash in nix-shell to
the bash in nix flake revealed that nix flake exposed "bash" in the PATH
but nix-shell exposed "bash-interactive". adding the dep makes the bash
kernel work on mac